requirements.txt
aioboto3==11.3.0
aiobotocore==2.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.5
aioitertools==0.12.0
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==3.7.1
attrs==24.2.0
bcrypt==4.0.1
boto3==1.28.17
botocore==1.31.17
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.104.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
jmespath==1.0.1
loguru==0.7.2
Mako==1.3.6
MarkupSafe==3.0.2
motor==3.6.0
multidict==6.1.0
openai==0.28.0
passlib==1.7.4
propcache==0.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
regex==2024.11.6
requests==2.32.3
rsa==4.9
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tiktoken==0.8.0
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.24.0
wrapt==1.16.0
yarl==1.17.2
